Refactor again
Sort components by type, put test file back in lib
